# specs/003-resolve-critical-issues/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: GASTAT International Dossier System API
  description: API for managing international relations, documents, and intelligence data
  version: 1.0.0
  contact:
    name: GASTAT Development Team
    email: dev@gastat.gov.sa
servers:
  - url: https://api.gastat.gov.sa/v1
    description: Production server
  - url: https://staging-api.gastat.gov.sa/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /organizations:
    get:
      summary: List organizations
      description: Get paginated list of organizations with filtering
      tags: [Organizations]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [government, ngo, international, private]
        - name: country_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended]
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create organization
      description: Create a new organization
      tags: [Organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'

  # MoUs Endpoints
  /mous:
    get:
      summary: List MoUs
      description: Get paginated list of MoUs with filtering
      tags: [MoUs]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, internal_review, external_review, negotiation, signed, active, renewed, expired]
        - name: organization_id
          in: query
          schema:
            type: string
            format: uuid
        - name: country_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of MoUs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MoU'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create MoU
      description: Create a new MoU
      tags: [MoUs]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateMoURequest'
      responses:
        '201':
          description: MoU created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoU'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [government, ngo, international, private]
        country_id:
          type: string
          format: uuid
        parent_organization_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, inactive, suspended]
        description:
          type: string
        website:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateOrganizationRequest:
      type: object
      required: [name, type, country_id]
      properties:
        name:
          type: string
          maxLength: 255
        type:
          type: string
          enum: [government, ngo, international, private]
        country_id:
          type: string
          format: uuid
        parent_organization_id:
          type: string
          format: uuid
        description:
          type: string
        website:
          type: string
          format: uri
        status:
          type: string
          enum: [active, inactive, suspended]
          default: active
    MoU:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        version:
          type: integer
        status:
          type: string
          enum: [draft, internal_review, external_review, negotiation, signed, active, renewed, expired]
        organization_id:
          type: string
          format: uuid
        country_id:
          type: string
          format: uuid
        document_path:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        effective_date:
          type: string
          format: date
        expiry_date:
          type: string
          format: date
        description:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateMoURequest:
      type: object
      required: [title, organization_id, country_id]
      properties:
        title:
          type: string
          maxLength: 255
        organization_id:
          type: string
          format: uuid
        country_id:
          type: string
          format: uuid
        description:
          type: string
        effective_date:
          type: string
          format: date
        expiry_date:
          type: string
          format: date
        file:
          type: string
          format: binary
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_prev:
          type: boolean
...